+++ lib/ui/BrowsePanel.dart
@@ -24,6 +24,7 @@ import 'package:illinois/ui/academics/StudentCourses.dart';
 import 'package:illinois/ui/athletics/AthleticsHomePanel.dart';
 import 'package:illinois/ui/canvas/CanvasCoursesListPanel.dart';
 import 'package:illinois/ui/canvas/GiesCanvasCoursesListPanel.dart';
+import 'package:illinois/ui/career/CareerPlanningLinks.dart';
 import 'package:illinois/ui/groups/GroupHome2Panel.dart';
 import 'package:illinois/ui/illini/WordlePanel.dart';
 import 'package:illinois/ui/messages/MessagesHomePanel.dart';
@@ -45,6 +46,7 @@ import 'package:illinois/ui/research/ResearchProjectsHomePanel.dart';
 import 'package:illinois/ui/safety/SafetyHomePanel.dart';
 import 'package:illinois/ui/gbv/GBVPathwaysPanel.dart';
 import 'package:illinois/ui/wellness/WellnessHomePanel.dart';
+import 'package:illinois/ui/wellness/WellnessLinksPanel.dart';
 import 'package:illinois/ui/widgets/FavoriteButton.dart';
 import 'package:illinois/ui/widgets/HeaderBar.dart';
 import 'package:illinois/utils/AppUtils.dart';
@@ -540,6 +542,7 @@ class _BrowseEntry extends StatelessWidget {
     'academics.my_illini'                  : 'external-link',
     'academics.due_date_catalog'           : 'external-link',
     'career_exploration.interest_explorer' : 'external-link',
+    'career_exploration.job_board'         : 'external-link',
     'music_and_news.daily_illini'          : 'external-link',
   };
 
@@ -583,10 +586,12 @@ class _BrowseEntry extends StatelessWidget {
       case "campus_guide.campus_highlights": _onTapCampusHighlights(context); break;
       case "campus_guide.my_campus_guide":   _onTapMyCampusGuide(context); break;
 
+      case "career_exploration.career_planing_links": _onTapCareerPlaningLinks(context); break;
       case "career_exploration.interest_explorer": _onTapInterestExplorer(context); break;
       case "career_exploration.digital_card": _onTapDigitalCard(context); break;
       case "career_exploration.skills_self_evaluation":_onTapSkillSelfEvaluation(context); break;
       case "career_exploration.essential_skills_coach":_onTapEssentialSkillCoach(context); break;
+      case "career_exploration.job_board":   _onTapJobBoard(context); break;
 
       case "dining.dining":                  _onTapDining(context); break;
 
@@ -619,6 +624,7 @@ class _BrowseEntry extends StatelessWidget {
       case "wellness.wellness_todo":            _onTapWellnessToDo(context); break;
       case "wellness.my_appointments":          _onTapWellnessAppointments(context); break;
       case "wellness.wellness_tips":            _onTapWellnessTips(context); break;
+      case "wellness.wellness_links":           _onTapWellnessLinks(context); break;
       case "wellness.wellness_health_screener": _onTapWellnessHealthScreener(context); break;
       case "wellness.wellness_success_team":    _onTapWellnessSuccessTeam(context); break;
     }
@@ -735,9 +741,19 @@ class _BrowseEntry extends StatelessWidget {
     );
   }
 
+  static void _onTapCareerPlaningLinks(BuildContext context) {
+    Analytics().logSelect(target: "Career Planning Links");
+    Navigator.push(context, CupertinoPageRoute(builder: (context) => CareerPlanningLinksPanel()));
+  }
+
   static void _onTapInterestExplorer(BuildContext context) {
     Analytics().logSelect(target: "Interest Explorer");
     _launchUrl(context, Config().interestExplorerUrl);
+  }
+
+  static void _onTapJobBoard(BuildContext context) {
+    Analytics().logSelect(target: "Virtual Job Board");
+    _launchUrl(context, Config().jobBoardUrl);
   }
 
   static void _onTapDigitalCard(BuildContext context) {
@@ -886,6 +902,11 @@ class _BrowseEntry extends StatelessWidget {
   static void _onTapWellnessTips(BuildContext context) {
     Analytics().logSelect(target: "Wellness Daily Tips");
     Navigator.push(context, CupertinoPageRoute(builder: (context) => WellnessHomePanel(contentType: WellnessContentType.dailyTips,)));
+  }
+
+  static void _onTapWellnessLinks(BuildContext context) {
+    Analytics().logSelect(target: "24/7 Hotlines & Links");
+    Navigator.push(context, CupertinoPageRoute(builder: (context) => WellnessLinksPanel()));
   }
 
   static void _onTapWellnessHealthScreener(BuildContext context) {